Move gatsby-theme-apollo to gatsby-theme-apollo-core
