+++ packages/gatsby-theme-apollo-core/gatsby-config.js
