+++ packages/gatsby-theme-apollo-core/index.js
