+++ packages/gatsby-theme-apollo-core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gatsby-theme-apollo",
+  "name": "gatsby-theme-apollo-core",
   "version": "1.1.0",
   "main": "index.js",
   "dependencies": {
