+++ packages/gatsby-theme-apollo-core/src/components/content-wrapper.js
