+++ packages/gatsby-theme-apollo-core/src/components/flex-wrapper.js
