+++ packages/gatsby-theme-apollo-core/src/components/header.js
