+++ packages/gatsby-theme-apollo-core/src/components/layout.js
