+++ packages/gatsby-theme-apollo-core/src/components/logo-title.js
