+++ packages/gatsby-theme-apollo-core/src/components/logo.js
