+++ packages/gatsby-theme-apollo-core/src/components/responsive-sidebar.js
