+++ packages/gatsby-theme-apollo-core/src/components/sidebar-nav/category.js
