+++ packages/gatsby-theme-apollo-core/src/components/sidebar-nav/index.js
