+++ packages/gatsby-theme-apollo-core/src/components/sidebar.js
