+++ packages/gatsby-theme-apollo-core/src/pages/404.js
