+++ packages/gatsby-theme-apollo-core/src/styles.less
