+++ packages/gatsby-theme-apollo-core/src/utils/colors.js
