+++ packages/gatsby-theme-apollo-core/static/favicon.ico
