+++ packages/gatsby-theme-apollo-docs/src/components/nav/index.js
@@ -2,7 +2,7 @@ import NavItem from './nav-item';
 import PropTypes from 'prop-types';
 import React from 'react';
 import styled from '@emotion/styled';
-import {breakpoints} from 'gatsby-theme-apollo';
+import {breakpoints} from 'gatsby-theme-apollo-core';
 import {formatValue} from '../../utils';
 
 const Container = styled.nav({
