+++ packages/gatsby-theme-apollo-docs/src/components/page-header.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import styled from '@emotion/styled';
-import {colors} from 'gatsby-theme-apollo';
+import {colors} from 'gatsby-theme-apollo-core';
 
 const Heading = styled.h1({
   ':not(:last-child)': {
