+++ packages/gatsby-theme-apollo-docs/src/components/typescript-api-box.js
@@ -6,7 +6,7 @@ import remark from 'remark';
 import remark2react from 'remark-react';
 import styled from '@emotion/styled';
 import withProps from 'recompose/withProps';
-import {colors, smallCaps} from 'gatsby-theme-apollo';
+import {colors, smallCaps} from 'gatsby-theme-apollo-core';
 
 const Container = styled.div({
   border: `1px solid ${colors.divider}`,
